Testing minecraft-like chunk sending

+++ src/main/java/ru/doh1221/wintymc/server/utils/location/Loc2D.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 @AllArgsConstructor
 public class Loc2D {
 
-    protected int x;
-    protected int z;
+    protected double x;
+    protected double z;
 
     public Loc2D() { }
